add: new ui, clean files and fix difficulty
