+++ styles/main.css
@@ -2,8 +2,18 @@
   box-sizing: border-box;
 }
 
+body {
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+  min-height: 100vh;
+  background-color: #131516 !important;
+}
+
 h1 {
   text-align: center;
+  color: #dd4814;
+  margin-top: 3rem;
 }
 
 table {
@@ -17,17 +27,18 @@ td.sudoku-case[data-value='0'] {
 }
 
 td {
-  border: 1px solid black;
+  border: 1px solid #8d8272;
   width: 40px;
   height: 40px;
   font-size: 24px;
   font-weight: bold;
-  background-color: #fff;
+  background-color: #191b1c;
+  color: #eee;
   cursor: pointer;
 }
 
 td:hover {
-  background-color: #eee;
+  background-color: #2c2f30;
 }
 
 .button-container {
@@ -40,9 +51,10 @@ td:hover {
   padding: 10px;
   font-size: 20px;
   font-weight: bold;
-  color: white;
+  color: #eee;
   border: none;
   cursor: pointer;
+  border-radius: 4px;
 }
 
 .easy-button {
@@ -64,9 +76,10 @@ td:hover {
   font-size: 20px;
   font-weight: bold;
   background-color: #4caf50;
-  color: white;
+  color: #eee;
   border: none;
   cursor: pointer;
+  border-radius: 4px;
 }
 
 .generate-button:hover {
@@ -83,38 +96,56 @@ td:hover {
   opacity: 0.8;
 }
 
+footer {
+  color: #eee;
+  padding: 20px;
+  text-align: center;
+  font-size: 14px;
+  margin-top: auto;
+  background-color: #191b1c !important;
+}
+
+footer a {
+  color: #dd4814;
+  text-decoration: none;
+}
+
+footer a:hover {
+  text-decoration: underline;
+}
+
 tr:nth-child(3n + 1) td:nth-child(3n + 1) {
-  border-top: 3px solid black;
-  border-left: 3px solid black;
+  border-top: 3px solid #8d8272;
+  border-left: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 1) td:nth-child(3n + 2) {
-  border-top: 3px solid black;
+  border-top: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 1) td:nth-child(3n + 3) {
-  border-top: 3px solid black;
-  border-right: 3px solid black;
+  border-top: 3px solid #8d8272;
+  border-right: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 2) td:nth-child(3n + 1) {
-  border-left: 3px solid black;
+  border-left: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 2) td:nth-child(3n + 3) {
-  border-right: 3px solid black;
+  border-right: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 3) td:nth-child(3n + 1) {
-  border-bottom: 3px solid black;
-  border-left: 3px solid black;
+  border-bottom: 3px solid #8d8272;
+  border-left: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 3) td:nth-child(3n + 2) {
-  border-bottom: 3px solid black;
+  border-bottom: 3px solid #8d8272;
 }
 
 tr:nth-child(3n + 3) td:nth-child(3n + 3) {
-  border-bottom: 3px solid black;
-  border-right: 3px solid black;
+  border-bottom: 3px solid #8d8272;
+  border-right: 3px solid #8d8272;
 }
