+++ templates/layout/footer.html
@@ -1,0 +1,4 @@
+<footer>
+  <p>Made in rust with Actix.</p>
+  <p>Copyright &copy; <a href="https://github.com/Jayllyz">Jayllyz</a></p>
+</footer>
