+++ templates/pages/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <title>{{ title }}</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" type="text/css" media="screen" href="/static/styles/main.css" />
+    <link rel="stylesheet" type="text/css" href="/static/styles/main.css" />
   </head>
   <body>
     <main>
@@ -31,5 +31,6 @@
         </form>
       </div>
     </main>
+    {% include "layout/footer.html" %}
   </body>
 </html>
